+++ noop/docs.go
@@ -1,0 +1,4 @@
+/*
+Package noop contains an implementation of the `gokv.Store` interface that does nothing.
+*/
+package noop
