+++ noop/go.mod
@@ -1,0 +1,8 @@
+module github.com/philippgille/gokv/noop
+
+go 1.18
+
+require (
+	github.com/philippgille/gokv v0.6.0
+	github.com/philippgille/gokv/util v0.6.0
+)
