--- nop/docs.go
@@ -1,4 +1,0 @@
-/*
-Package nop contains an implementation of the `gokv.Store` interface that does nothing.
-*/
-package nop
