--- nop/go.mod
@@ -1,11 +1,0 @@
-module github.com/philippgille/gokv/nop
-
-go 1.18
-
-require (
-	github.com/philippgille/gokv/util v0.6.0
-)
-
-require (
-	github.com/philippgille/gokv v0.6.0 // indirect
-)
